test: reorganize test structure
